Fix #5661, update react-webcam dependency (#5711)
- fix camera jitter bug camera_input repeatedly reopens webcam as window width changes (in wide layout) #5661
- update react-webcam to version 7.0.1.
We request video stream from a browser using the width prop, but it could have been changed very often (when we resize the window), which was leading to spamming the system queue of requests to the web camera. As a solution, we use debouncing width change handler and only request new video streams with constraints based on the width that stabilized after window resizes.