Fix detecting arrow df as widget when not editable (#6127)
The data editor introduced a new kind of state for an element, being a
widget only conditionally. Element replay code relied on the presence of
an id field in the protobuf for determining what is a widget, but arrow
dfs don't act like a widget, and don't call `register_widget`, unless
used with the data editor. In this situation, the id isn't actually set
on the protobuf, so we can use it being set for our widget detection.

Fixes #6103