Cache storage abstraction (#6052)
Introduce the new CacheStorage and CacheStorageManager abstractions for @st.cache_data. This allows customizing the way that cache values are stored within Streamlit.

RuntimeConfig now takes a CacheStorageManager interface, and cache_data_api constructs the CacheStorage using that instance.

We ship with a concrete LocalDiskCacheStorage and LocalDiskCacheStorageManager
implementation, which just stores cache items in memory and on disk if persist=True is passed to st.cache_data (as before). This is what server.py uses when it initializes its Runtime instance.

We also implemented InMemoryCacheStorageWrapper class, which could serve as a facade and wrap any CacheStorage with an in_memory first layer cache.

Also, for testing purposes, DummyCacheStorage and DummyCacheStorageManagerimplemented.

Because the LocalDiskCacheStorageManager instance exists when a Runtime has been created, scripts that run in "raw" mode (i.e. streamlit._is_running_with_streamlit == False) no longer have access to that manager from Runtime. So in get_storage_manager in cache_data_api we fall back to DummyCacheStorageManager.