Don't forget to stop script execution on websocket close (#6204)
There's a bug in the websocket reconnect logic that was implemented in #5856 where I forgot
to stop script execution in the websocket disconnect handler. This is fine with the vast majority
of scripts since most of them execute relatively quickly, but it does result in some weird behavior
with the (admittedly, rare and not-very-well-supported) use-case where a script is written to be
long-lived (via a while True loop or something similar). See #6166 for more information on the
bug.

This PR fixes this bug by simply requesting that the scriptrunner stop execution when on websocket
disconnect.