Enrich communication between parent and child iframe (#6342)
When iframing Streamlit, this PR makes it possible for the host frame to send signals to Streamlit app to STOP_SCRIPT, RERUN_SCRIPT and to CLEAR_CACHE.

In opposite direction (from Streamlit iframe to parent iframe) send messages whenever the "script run state" changes.

Note that this is not meant to be a public API

This PR also adds functionality to frontend/hostframe.html to test host-guest communication described here. See that file for usage.