Set default filename for if it missing in `st.download_button` widget. (#6832)
Previously in case of missing filename, we computed it based on the page title, and in order to have it in the endpoint we added the title query param to download_url.

This could lead to undesired results (for example in the case of AWS S3 pre-signed URL modification breaks the link). So this PR removes this logic completely and uses a generic filename instead when filename is not provided by a user.