Lazy-load emoji module to improve performance (#8109)
## Describe your changes

The emoji data is the biggest object when running a blank Streamlit app.
Compiling the regex is also slightly expensive. However, the emoji data
is only required if there is a check for emojis; many apps might not
require this. Therefore, this PR makes the emoji module to lazy load
only if it is actually required.

This also adds a precheck for emoji checks to make sure that the string
even contains non alphanumeric characters before using the more
expensive emoji regex.

## GitHub Issue Link (if applicable)

Related to https://github.com/streamlit/streamlit/issues/6066

## Testing Plan

- Added e2e test to check that some lazy-loaded modules are not imported
in an almost blank Streamlit app.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.