Make st.help more resilient with conditional members (#8228)
## Describe your changes

`st.help` leverages `dir(object)` to determine the members in the object
to expose to the docstring. If an object makes a member conditional
based on internal state, `st.help` can throw an `AttributeError`.
Ideally, in these cases, the developer of the object would implement the
`__dir__` method to handle conditional members, but Streamlit can be
more resilient in handling this situation.

The example that brought this up was in sklearn, and I created
https://github.com/scikit-learn/scikit-learn/issues/28558 with them to
better handle this scenario.

## Testing Plan

- Unit Tests should cover the change with the member available and
unavailable

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.