Fix flaky e2e test (#8232)
## Describe your changes

There is a strange error causing flakiness in a combination of dataframe
with use_container_width in popover when running in Chromium within our
CI. I cannot reproduce this locally and only have very limited
information about this issue. Not sure how specific this is to our
Playwright testing setup.

This PR will remove this flakiness, but I will continue to investigate
the underlying issue.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.