Allow custom themes to override embed options query parameter (#8021)
## Describe your changes

We set the theme based on the query param each time the script run is
finished. This can break other theme settings, so we want to respect the
query parameters only on initial load (in determining the default theme.

## GitHub Issue Link (if applicable)
Closes #7118 

## Testing Plan

- Updated JS Unit Tests for the default theme. There were no tests for
the script finished action, so I did not include tests for the removal.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.