Update nightly release to OpenID Connect (#8460)
## Describe your changes

We are setting up changes to make sure our releases are more secure. We
are starting with the nightly release removing the need to use a
username/password and setting up a Trusted Publisher on PyPi. These are
the changes needed based on [this
resource](https://docs.github.com/en/actions/deployment/security-hardening-your-deployments/configuring-openid-connect-in-pypi).

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.