Migrate st.heading cypress test to playwright (#8598)
## Describe your changes

This PR migrates the existing st_heading and st_subheading tests from
Cypress to playwright. Since the two tests touch the same area, they are
merged into one test script.

The migration happens in preparation of the
https://github.com/streamlit/streamlit/pull/8587 PR. In that PR a couple
of screenshots need to be updated but I don't want to touch the legacy
cypress test artifacts.

## GitHub Issue Link (if applicable)

## Testing Plan

- ✅ E2E Tests

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.