Allow protobuf v5 (#8624)
## Describe your changes
Increase upper-bound on protobuf to allow v5. It seems like protobuf <5
is not compatible with Python 3.12, so this bump helps streamlit's
compatibility with Python 3.12.

## GitHub Issue Link (if applicable)

## Testing Plan

No additional tests are required because this is just allowing a
dependency bump. The current tests in the repository that run at CI time
should be a sufficient verification.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.

Co-authored-by: Lukas Masuch <Lukas.Masuch@gmail.com>