Add internal config to indicate that its running as e2e test (#8746)
## Describe your changes

Add a new internal config option `global.e2eTest` that indicates that
the app is running as part of e2e tests. This is used to activate
certain internal functionality that we use for advanced e2e testing.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.