Migrate streamlit hello to MPAv2 (#8806)
## Describe your changes

This PR converts `streamlit hello` to use Multipage Apps v2. We overall
kept the change very simple to mimic the same look and feel. We can
adjust it accordingly.

A minor change to improve types is here when we cannot find the context.
This should be impossible in a Streamlit app, but for bare execution and
to make the return type to ALWAYS be a Page, we return the default page
to run.

## Testing Plan

- Same E2E and unit tests should pass

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.