Treat @streamlit as an internal import for eslint (#9191)
## Describe your changes

This PR lets eslint know that `@streamlit` is an internal package and
that it should sort it as such. I found it a bit odd that this rule
changed ZERO things in lib... but from my manual verification, it seems
okay..?

Follow up from this discussion:
https://github.com/streamlit/streamlit/pull/9153#discussion_r1695996337


---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.