Refactor integration test setup (#9263)
## Describe your changes

Refactors the snowflake integration test workflow to a more generic
integration test that allows the testing of any optional
dependencies/integrations. This also moves a couple of tests from our
core unit test workflow to the integration test workflow:

- Snowflake integrations
- SQL-connection (uses SQLalchemy)
- Dataframe libraries Integrations
- Langchain Handler 

This makes it easier for us to add tests for more integrations since it
will not interfere with our core tests.

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.