Fix nested fragment stale elements (#9246)
## Describe your changes

Closes https://github.com/streamlit/streamlit/issues/9233
Closes https://github.com/streamlit/streamlit/issues/9267

This PR changes the way how we deprecate child nodes in fragment runs. 
Right now, we have the issue that child fragment elements can be stale
even if the parent changed.

The PR changes the behavior now as following:
- when the parent fragment was changed during a run, we consider all
_nested_ fragment elements as stale as well unless they were also
modified during the same run (the scriptRunId is correct).

---

## Testing Plan

- Explanation of why no additional tests are needed
- Unit Tests (JS and/or Python)
- Add new frontend test to ensure that stale nested fragments are
removed
- E2E Tests
  - Update the existing `st_fragments_nested` test.
- Any manual testing needed?

---

**Contribution License Agreement**

By submitting this pull request you agree that all contributions to this
project are made under the Apache 2.0 license.